pyproject.toml
[project]
name = "learn-video"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.3",
    "mediapipe>=0.10.21",
    "opencv-python>=4.11.0.86",
    "pillow>=11.2.1",
]
